LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/perspective-origin-expected.txt

CONSOLE MESSAGE: Error: '<position>' is not a valid CSS component

Harness Error (FAIL), message = Error: '<position>' is not a valid CSS component

FAIL Can set 'perspective-origin' to CSS-wide keywords: initial assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'perspective-origin' to CSS-wide keywords: inherit assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'perspective-origin' to CSS-wide keywords: unset assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'perspective-origin' to CSS-wide keywords: revert assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'perspective-origin' to var() references:  var(--A) assert_equals: expected 2 but got 1
FAIL Can set 'perspective-origin' to the 'none' keyword: none Bad value for shorthand CSS property
